# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/payments/paystack/initialize": {
      "post": {
        "operationId": "PaymentsController_initializePayment_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitializePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/paystack/verify/{reference}": {
      "get": {
        "operationId": "PaymentsController_verifyPayment_v1",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/paystack/charge/authorize": {
      "post": {
        "operationId": "PaymentsController_authorizeCharge_v1",
        "parameters": [],
        "responses": {
          "501": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/paystack/transfer/initialize": {
      "post": {
        "operationId": "PaymentsController_initializeTransfer_v1",
        "parameters": [],
        "responses": {
          "501": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/paystack/transfer/status/{ref}": {
      "get": {
        "operationId": "PaymentsController_getTransferStatus_v1",
        "parameters": [],
        "responses": {
          "501": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/paystack/refund": {
      "post": {
        "operationId": "PaymentsController_initializeRefund_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/stripe/payment-intent": {
      "post": {
        "operationId": "StripeController_createPaymentIntent_v1",
        "parameters": [],
        "responses": {
          "501": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/stripe/payment-intent/{id}/confirm": {
      "post": {
        "operationId": "StripeController_confirmPaymentIntent_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/stripe/setup-intent": {
      "post": {
        "operationId": "StripeController_createSetupIntent_v1",
        "parameters": [],
        "responses": {
          "501": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/stripe/refund": {
      "post": {
        "operationId": "StripeController_issueRefund_v1",
        "parameters": [],
        "responses": {
          "501": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/stripe/payment-methods": {
      "get": {
        "operationId": "StripeController_listPaymentMethods_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/stripe/payment-methods/{id}": {
      "delete": {
        "operationId": "StripeController_detachPaymentMethod_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
...
    "/payments/pod/eligibility": {
      "get": {
        "operationId": "PodController_checkEligibility_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
    "/payments/pod/confirm": {
      "post": {
        "operationId": "PodController_confirmPod_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cadna Mart API",
    "description": "Backend API for the Cadna Mart marketplace.\n\n### Response envelope\nEvery successful response is wrapped by the global response interceptor:\n```json\n{\n  \"success\": true,\n  \"statusCode\": 200,\n  \"data\": { ... },\n  \"message\": \"...\",\n  \"meta\": { \"correlationId\": \"...\", \"timestamp\": \"...\" }\n}\n```\nWhen documenting individual endpoints below, the schema describes the `data` payload only.\n\n### Authentication\n- **Bearer JWT** (`Authorization: Bearer <accessToken>`) for authenticated user routes.\n- **Guest token** (`x-guest-token: <token>`) for guest-cart endpoints. Returned once from `POST /cart`.\n\n### Error shape\nErrors flow through the global `AllExceptionsFilter`:\n```json\n{\n  \"success\": false,\n  \"statusCode\": 400,\n  \"message\": \"Validation failed\",\n  \"errorCode\": \"BAD_REQUEST\",\n  \"path\": \"/api/v1/...\",\n  \"correlationId\": \"...\"\n}\n```",
    "version": "1.0",
    "contact": {
      "name": "Cadna Mart Engineering",
      "url": "",
      "email": "cadnamart@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "https://cadna-mart-be-nsz2.onrender.com/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned from /auth/login or /auth/otp/verify"
      },
      "guestToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-guest-token",
        "description": "Guest cart access token (returned once from POST /cart for guests)"
      }
    },
    "schemas": {
...
      "InitializePaymentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}